// backstage/src/models/interfaces/UrlParams.ts
export default interface UrlParams {
  eventId?: string;
  seasonId?: string;
  ticketHoldId?: string;
  ticketTypeId?: string;
  upgradeTypeId?: string;
  promotionId?: string;
  customFieldId?: string;
  orderId?: string;
  venueId?: string;
  artistId?: string;
  orgId?: string;
  customerId?: string;
  roleId?: string;
  feeId?: string;
  chartKey?: string;
  modalType?: string;
  onboarding?: string;
}